requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-chunked-upload==2.0.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
multidict==6.0.4
openai==0.28.1
packaging==23.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
yarl==1.9.2
